requirements.txt
matplotlib==3.7.1
numpy==1.24.1
opencv-python==4.7.0.72
pandas==2.0.2
pillow==9.3.0
torch==2.0.1+cu117
torchvision==0.15.2+cu117
tqdm==4.65.0

